+++ pakyow-realtime/spec/features/installing_spec.rb
@@ -9,7 +9,7 @@ RSpec.describe "installing the socket" do
   it "installs the socket" do
     expect(call("/")[2]).to include_sans_whitespace(
       <<~HTML
-        <meta name="pw-socket" data-ui="socket(global: true, endpoint: ws://localhost/pw-socket?id=MTIzMjE=--FGhnpS-4JBlFz4V-78zKBAIr0m7e-Mf1mryud9JZt0U=)">
+        <meta name="pw-socket" data-ui="socket(global: true, endpoint: ws://localhost/pw-socket?id=MTIzMjE=~FGhnpS-4JBlFz4V-78zKBAIr0m7e-Mf1mryud9JZt0U=)">
       HTML
     )
   end
@@ -22,7 +22,7 @@ RSpec.describe "installing the socket" do
     it "configures the socket to connect directly to the app" do
       expect(call("/")[2]).to include_sans_whitespace(
         <<~HTML
-          <meta name="pw-socket" data-ui="socket(global: true, endpoint: ws://localhost:3000/pw-socket?id=MTIzMjE=--FGhnpS-4JBlFz4V-78zKBAIr0m7e-Mf1mryud9JZt0U=)">
+          <meta name="pw-socket" data-ui="socket(global: true, endpoint: ws://localhost:3000/pw-socket?id=MTIzMjE=~FGhnpS-4JBlFz4V-78zKBAIr0m7e-Mf1mryud9JZt0U=)">
         HTML
       )
     end
@@ -32,7 +32,7 @@ RSpec.describe "installing the socket" do
     it "configures the socket to connect with wss" do
       expect(call("/", scheme: "https")[2]).to include_sans_whitespace(
         <<~HTML
-          <meta name="pw-socket" data-ui="socket(global: true, endpoint: wss://localhost/pw-socket?id=MTIzMjE=--FGhnpS-4JBlFz4V-78zKBAIr0m7e-Mf1mryud9JZt0U=)">
+          <meta name="pw-socket" data-ui="socket(global: true, endpoint: wss://localhost/pw-socket?id=MTIzMjE=~FGhnpS-4JBlFz4V-78zKBAIr0m7e-Mf1mryud9JZt0U=)">
         HTML
       )
     end