+++ pakyow-support/CHANGELOG.md
@@ -3,6 +3,7 @@
   * [changed] `Config` now supports multiple default config blocks per environnment
   * [changed] `Config#defaults` now supports passing multiple environments
   * [fixed] Bug that occurs if a named action shares a name with a method on the pipelined object
+  * [fixed] Message verification no longer fails when the digest contains `--`
 
 # 1.0